requirements.txt
certifi==2024.8.30
googleapis-common-protos==1.65.0
grpcio==1.67.0
lz4==4.3.3
pinecone==5.3.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
